meenfee_dev/requirements_meenfee_4_oct.txt
apns2==0.4.1
asn1crypto==0.24.0
astroid==2.1.0
authy==2.2.3
Babel==2.6.0
beautifulsoup4==4.8.0
cachetools==3.1.0
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
cryptography==2.5
defusedxml==0.5.0
Django==2.2.1
django-admin-bootstrapped==2.5.7
django-allauth==0.38.0
django-bootstrap3==11.0.0
django-common-helpers==0.9.2
django-cors-headers==2.5.2
django-cron==0.5.1
django-filter==2.1.0
django-map-widgets==0.2.2
django-phonenumber-field==2.1.0
django-rest-auth==0.9.3
django-tinymce==2.8.0
django-tinymce4-lite==1.7.5
django-twilio==0.9.3
django-uuidfield==0.5.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
drf-rw-serializers==1.0.1
drf-writable-nested==0.5.1
fcm-django==0.2.21
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
gunicorn==19.9.0
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
httplib2==0.12.3
hyper==0.7.0
hyperframe==3.2.0
idna==2.8
isort==4.3.4
jsmin==2.2.2
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauth2client==4.1.3
oauthlib==3.0.1
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
py-vapid==1.4.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pybase64==0.5.0
pycparser==2.19
pycrypto==2.6.1
pyfcm==1.4.5
PyJWT==1.7.1
pylint==2.2.2
PySocks==1.6.8
python3-openid==3.1.0
pytz==2018.9
pywebpush==1.8.0
requests==2.21.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
rsa==4.0
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.2
South==1.0.2
sqlparse==0.3.0
twilio==6.23.1
typed-ast==1.2.0
uritemplate==3.0.0
urllib3==1.24.1
wrapt==1.11.1
yawd-admin==0.7.0
